% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk.R
\name{AverageReplicates}
\alias{AverageReplicates}
\title{Calculate average distance given a set of replicate distances}
\usage{
AverageReplicates(dist.matrix)
}
\description{
Calculate average distance given a set of replicate distances
}
