% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam.R
\name{bam2bw}
\alias{bam2bw}
\title{Convert BAM to bigwig}
\usage{
bam2bw(
  file,
  bw.path = gsub(pattern = "\\\\.bam$", replacement = ".bw", x = file, ignore.case =
    TRUE),
  paired = FALSE,
  strand = "both",
  method = c("native", "deeptools"),
  bamcoverage.path = NULL,
  normalization.method = NULL,
  cores = 1L
)
}
\description{
Convert BAM to bigwig
}
