% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqdist.R
\name{CollapseKmerCounts}
\alias{CollapseKmerCounts}
\title{Collpase kmer data for kmers that are within a hamming distance of 1.
This method is based on collapse_polyN function from Parsebio's splitpipe pipeline.}
\usage{
CollapseKmerCounts(kmers, counts)
}
\arguments{
\item{kmers}{A vector of kmers}

\item{counts}{A vecotr of counts associated with kmers}
}
\description{
Collpase kmer data for kmers that are within a hamming distance of 1.
This method is based on collapse_polyN function from Parsebio's splitpipe pipeline.
}
