% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_predictions.R
\name{ComparePredictions}
\alias{ComparePredictions}
\title{Compare accuracy of prediction of two different annotation methods}
\usage{
ComparePredictions(
  object,
  celltype.1,
  celltype.2,
  predcol.1,
  predcol.2,
  scorecol.1,
  scorecol.2,
  assay = "SCT",
  slot = "data",
  use.presto = FALSE,
  use.fastglm = FALSE,
  do.plot = FALSE
)
}
\arguments{
\item{object}{Seurat object}

\item{celltype.1}{Celltype 1}

\item{celltype.2}{Celltype 2}

\item{predcol.1}{Metadata column name for first method's prediction scores}

\item{predcol.2}{Metadata column name for second methods' prediction scores}

\item{use_scale_data_for_glm}{Whether to use scaled data for running logistic regression}
}
\description{
This method uses consensus prediction between the methods as a ground truth to find markers for
each of the two celltypes where the two methods have a consensus and then uses these markers
to predict the most likely celltype of cells where the predictions differ
}
