% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{ConfusionHeatmap}
\alias{ConfusionHeatmap}
\title{Plot confusion heatmap for comparing the overlap between two sets of predictions}
\usage{
ConfusionHeatmap(predictions.1, predictions.2)
}
\description{
Plot confusion heatmap for comparing the overlap between two sets of predictions
}
