% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_helper.R
\name{convert_to_ratio}
\alias{convert_to_ratio}
\title{Convert text string to ratio}
\usage{
convert_to_ratio(x)
}
\arguments{
\item{x}{String}
}
\value{
ratio Float
}
\description{
Convert text string to ratio
}
