% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam.R
\name{CountReadsinParsebioBam}
\alias{CountReadsinParsebioBam}
\title{Count reads in a parsebio output bam}
\usage{
CountReadsinParsebioBam(file, name.sep = "__", readtype.num = 2, cells = NULL)
}
\description{
Count reads in a parsebio output bam
}
