% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk.R
\name{CreateKPseudoBulk}
\alias{CreateKPseudoBulk}
\title{Create a pseudobulk of any seurat object with replicates}
\usage{
CreateKPseudoBulk(
  object,
  ident.name = "subclass.predicted.id",
  k_replicates = 3,
  seed = 42
)
}
\description{
Create a pseudobulk of any seurat object with replicates
}
