% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{DoPCA}
\alias{DoPCA}
\title{Perform PCA}
\usage{
DoPCA(
  mtx.genebycell,
  npcs = 50,
  weight.by.var = TRUE,
  do.scale = FALSE,
  do.center = TRUE,
  method = "prcomp_irlba"
)
}
\description{
Perform PCA
}
