% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atac.R
\name{DoTn5Correction}
\alias{DoTn5Correction}
\title{Perform correction for Tn5 cut site
Increase forward mapping start positions by 4bp
Decrease reverse mapping end positions by 5bp.}
\usage{
DoTn5Correction(df)
}
\arguments{
\item{df}{A bed red in as a dataframe}
}
\description{
Perform correction for Tn5 cut site
Increase forward mapping start positions by 4bp
Decrease reverse mapping end positions by 5bp.
}
