% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbmodel.R
\name{DropoutPlot}
\alias{DropoutPlot}
\title{Plot expected vs observed dropouts (zero observation)}
\usage{
DropoutPlot(counts = NULL, annotate = FALSE, logx = TRUE, linewidth = 0.75)
}
\description{
Plot expected vs observed dropouts (zero observation)
}
