% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{FetchGEOFiles}
\alias{FetchGEOFiles}
\title{Fetch supplemetary files from GEO}
\usage{
FetchGEOFiles(gse, download.dir = getwd(), download.files = FALSE, ...)
}
\description{
Fetch supplemetary files from GEO
}
