% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{FormatEnrichrOutput}
\alias{FormatEnrichrOutput}
\title{Format GO output of EnrichR}
\usage{
FormatEnrichrOutput(go)
}
\arguments{
\item{go}{Dataframe as returned by running enrichR}
}
\value{
df Formatted dataframe
}
\description{
Format GO output of EnrichR
}
