% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{GetSequencesFromGenome}
\alias{GetSequencesFromGenome}
\title{Fetch sequences from a DNAStringSet object}
\usage{
GetSequencesFromGenome(genomeseq, gr)
}
\arguments{
\item{genomeseq}{A DNAStringSet object}

\item{gr}{A granges object}
}
\description{
Fetch sequences from a DNAStringSet object
}
