% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{GiniIndex}
\alias{GiniIndex}
\title{Calculate gini index to quantify the specificity of a gene to a cluster}
\usage{
GiniIndex(x, weights = NULL)
}
\arguments{
\item{x}{A vector of gene's normalized expression}

\item{weights}{A vector of weights for observations (default is 1)}
}
\value{
A scalar between 0-1 where 1 indicates ubiquitous expression.
}
\description{
Calculate gini index to quantify the specificity of a gene to a cluster
}
