% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqdist.R
\name{HammingDistance}
\alias{HammingDistance}
\title{Find the hamming distance between two strings}
\usage{
HammingDistance(string1, string2)
}
\arguments{
\item{input1}{String}

\item{input2}{String}
}
\description{
Find the hamming distance between two strings
}
