% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbmodel.R
\name{MeanVarFit}
\alias{MeanVarFit}
\title{Perform mean-variance calculation on the count matrix.}
\usage{
MeanVarFit(counts = NULL, means = NULL, variance = NULL)
}
\description{
Perform mean-variance calculation on the count matrix.
}
