% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MeanVarPlot}
\alias{MeanVarPlot}
\title{Plot mean variance relationship for counts or dataframe}
\usage{
MeanVarPlot(counts = NULL, df = NULL, annotate = FALSE, logxy = F)
}
\description{
Plot mean variance relationship for counts or dataframe
}
