% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.R
\name{NZstatsToZstats}
\alias{NZstatsToZstats}
\title{Convert mean and variance calculated on non-zero entries into
mean and variance if zeroes were to be included. This method is
useful when mean and variance are calculated for features in a data
frame where the zero entries have been dropped such as when using the
\code{SparseMatrixToDF} method.}
\usage{
NZstatsToZstats(nzmean, nzvariance, n, z)
}
\arguments{
\item{nzmean}{Mean of non-zero entries}

\item{n}{Number of non-zero entries}

\item{z}{Number of zero entries}

\item{nzvar}{Mean of non-zero entries}
}
\description{
Convert mean and variance calculated on non-zero entries into
mean and variance if zeroes were to be included. This method is
useful when mean and variance are calculated for features in a data
frame where the zero entries have been dropped such as when using the
\code{SparseMatrixToDF} method.
}
