% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{PerGeneCorrelation}
\alias{PerGeneCorrelation}
\title{Get correlation per gene given two matrices (genes as rows, celltypes as columns)}
\usage{
PerGeneCorrelation(mat1, mat2)
}
\arguments{
\item{mat1}{Normalized matrix with genes as rows, cells as columns}

\item{mat2}{Normalized matrix with genes as rows, cells as columns}
}
\value{
A named list of genewise correlations.
}
\description{
Get correlation per gene given two matrices (genes as rows, celltypes as columns)
}
