% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{plot_GO}
\alias{plot_GO}
\title{Plot Gene ontology terms on a dotplot}
\usage{
plot_GO(go, title)
}
\arguments{
\item{go}{Formatted data frame}

\item{title}{Title of the plot}
}
\description{
Plot Gene ontology terms on a dotplot
}
