% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_predictions.R
\name{PredictionComparisonHeatmap}
\alias{PredictionComparisonHeatmap}
\title{Plot confusion matrices for prediction}
\usage{
PredictionComparisonHeatmap(predictions.1, predictions.2)
}
\description{
Plot confusion matrices for prediction
}
