% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{PredictPCA}
\alias{PredictPCA}
\title{Calculate reconstruction error for a new dataset given feature loadings}
\usage{
PredictPCA(feature.loadings, newdata)
}
\description{
Calculate reconstruction error for a new dataset given feature loadings
}
