% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ProjectDataLongVector}
\alias{ProjectDataLongVector}
\title{Project a given matrix along a vector
The vector is first standardized}
\usage{
ProjectDataLongVector(mtx.genebycell, f.vector)
}
\description{
Project a given matrix along a vector
The vector is first standardized
}
