% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{PseudoBulkHeatmap}
\alias{PseudoBulkHeatmap}
\title{Pseudu bulk heatmap}
\usage{
PseudoBulkHeatmap(
  object,
  aggregate.by,
  features.to.plot,
  assay = "SCT",
  cellheight = 20,
  cellwidth = 20,
  use.scaledata = FALSE,
  idents.to.plot = NULL
)
}
\description{
Pseudu bulk heatmap
}
