% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenx.R
\name{Read10X_molinfo}
\alias{Read10X_molinfo}
\title{Read 10X molecular_info.h5 file}
\usage{
Read10X_molinfo(
  filename,
  version.10x = NULL,
  make.umi.seq = FALSE,
  return.raw = FALSE
)
}
\arguments{
\item{filename}{Path to molecular_info.h5}

\item{make.umi.seq}{Whether to convert the integer umi values to a sequence (post 2-bit encoding/decoding)}

\item{return.raw}{Whether a raw value or just the read count summary at the cell barcode level}

\item{version.10X}{"v2" or "v3"}
}
\description{
Read 10X molecular_info.h5 file
}
