% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{Read10XCustom}
\alias{Read10XCustom}
\title{Modified function to support differente filenames}
\usage{
Read10XCustom(
  data.dir = NULL,
  gene.column = 2,
  unique.features = TRUE,
  strip.suffix = FALSE,
  barcode_filename = "barcodes.tsv",
  gene_filename = "genes.tsv",
  matrix_filename = "matrix.mtx"
)
}
\description{
Modified function to support differente filenames
}
