% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{ReadKallistoOutput}
\alias{ReadKallistoOutput}
\title{Read raw output from kallisto/bustools}
\usage{
ReadKallistoOutput(barcodes_file, genes_file, mtx_file)
}
\arguments{
\item{barcodes_file}{Location of bustools output.}
}
\value{
list containing spliced, unspliced and merged counts
}
\description{
Read raw output from kallisto/bustools
}
