% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{ReadKBSpliced}
\alias{ReadKBSpliced}
\title{Read output from spliced mode of kallisto/bustools}
\usage{
ReadKBSpliced(base.dir)
}
\arguments{
\item{base.dir}{Location of bustools output.}
}
\value{
list containing spliced, unspliced and merged counts
}
\description{
Read output from spliced mode of kallisto/bustools
}
