% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{ReadParsebioOutput}
\alias{ReadParsebioOutput}
\title{Read output from parsebio pipeline}
\usage{
ReadParsebioOutput(
  path,
  add.hexR.assay = TRUE,
  add.polyT.assay = TRUE,
  verbose = TRUE
)
}
\description{
Read output from parsebio pipeline
}
