% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ReconErrorPCA}
\alias{ReconErrorPCA}
\title{Perform PCA and return loadings along with reconstruction error}
\usage{
ReconErrorPCA(original.data, npcs = 50, weight.by.var = TRUE)
}
\description{
Perform PCA and return loadings along with reconstruction error
}
