% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{sct_vs_std}
\alias{sct_vs_std}
\title{SCT vs STD workflow comparison}
\usage{
sct_vs_std(
  seu,
  dims = 1:30,
  nfeatures = 3000,
  idents = c("seurat_clusters"),
  outdir = NULL,
  overwrite = TRUE
)
}
\arguments{
\item{seu}{Seurat object}

\item{dims}{Number of dimensions for PCA/UMAP}

\item{nfeatures}{Number of variable features for STD workflow}
}
\description{
SCT vs STD workflow comparison
}
