% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sct.R
\name{SCTReconstructionError}
\alias{SCTReconstructionError}
\title{SCT's reconstruction error}
\usage{
SCTReconstructionError(
  seu,
  dims = 1:30,
  nfeatures = 3000,
  idents = c("seurat_clusters"),
  outdir = NULL,
  overwrite = TRUE
)
}
\arguments{
\item{seu}{Seurat object}
}
\description{
SCT's reconstruction error
}
