% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqdist.R
\name{SingleMutations}
\alias{SingleMutations}
\title{Return all possible 1bp mutated sequence}
\usage{
SingleMutations(seq)
}
\arguments{
\item{seq}{String}
}
\value{
A vector of mutated sequences
}
\description{
Return all possible 1bp mutated sequence
}
