% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.R
\name{SparseSpearmanCor}
\alias{SparseSpearmanCor}
\title{Calculate spearman correlation on sparse matrix(ces)}
\usage{
SparseSpearmanCor(X, Y = NULL, cov = FALSE)
}
\description{
Calculate spearman correlation on sparse matrix(ces)
}
