% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{SparsifyParseDataframe}
\alias{SparsifyParseDataframe}
\title{Convert parsebio's long dataframe to a sparse matrix}
\usage{
SparsifyParseDataframe(df, all_genes, all_cells)
}
\description{
Convert parsebio's long dataframe to a sparse matrix
}
