% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam.R
\name{SplitParsebioBam}
\alias{SplitParsebioBam}
\title{Split parsebio bam based on HexR or PolyT priming}
\usage{
SplitParsebioBam(
  file,
  hexR.out = gsub(pattern = "\\\\.bam$", replacement = "_hexR.bam", x = file, ignore.case
    = TRUE),
  polyT.out = gsub(pattern = "\\\\.bam$", replacement = "_polyT.bam", x = file,
    ignore.case = TRUE),
  name.sep = "__",
  readtype.num = 2,
  metadata = NULL,
  group.by = NULL
)
}
\description{
Split parsebio bam based on HexR or PolyT priming
}
