% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftover.R
\name{SummarisePeakOverlap}
\alias{SummarisePeakOverlap}
\title{importFrom GenomicRanges findOverlaps}
\usage{
SummarisePeakOverlap(
  query,
  subject,
  query_name = "query",
  subject_name = "subject",
  query_peak_col = "peak",
  subject_peak_col = "peak"
)
}
\description{
importFrom GenomicRanges findOverlaps
}
