% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_predictions.R
\name{SummarisePredDiffs}
\alias{SummarisePredDiffs}
\title{Summarize difference in prediction obtained from two prediction methods}
\usage{
SummarisePredDiffs(
  object,
  predcol.1,
  predcol.2,
  min.diff.cells = 3,
  diff.only = FALSE
)
}
\arguments{
\item{object}{Seurat object}

\item{predcol.1}{Metadata column name for first method's predictions}

\item{predcol.2}{Metadata column name for second methods' predictions}

\item{min.diff.cells}{Return only comparisons where the difference between two methods exceeds this. Default is 3}

\item{diff.only}{Return statistics only for cases where the predictions differ. Default is false which will return all comparisons}
}
\description{
Summarize difference in prediction obtained from two prediction methods
}
