% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{TauIndex}
\alias{TauIndex}
\title{Calculate Tau index to quantify the specificity of a gene to a cluster}
\usage{
TauIndex(x)
}
\arguments{
\item{x}{A vector of gene's normalized expression}
}
\value{
A scalar between 0-1 where 1 indicates ubiquitous expression.
}
\description{
Calculate Tau index to quantify the specificity of a gene to a cluster
}
