% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{TFIDFMarkers}
\alias{TFIDFMarkers}
\title{Run DE on counts based on TFIDF}
\usage{
TFIDFMarkers(counts, clusters, expression.cutoff = NULL)
}
\arguments{
\item{counts}{A dgCMatrix with raw or sequencing depth corrected counts}

\item{clusters}{A character vector of assigned clusters corresponding
to columns in the counts matrix}

\item{expression.cutoff}{Cutoff to determine whether a gene is expressed or not. Default is NULL
which uses the median count}
}
\description{
Run DE on counts based on TFIDF
}
