% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenx.R
\name{TwoBitToSeq}
\alias{TwoBitToSeq}
\title{Convert a given integer to 2-bit and eventually ATCG..
sequence
Based on https://kb.10xgenomics.com/hc/en-us/articles/360004105372-How-do-you-decompress-the-2-bit-barcode-sequences-in-molecule-info-h5-file-}
\usage{
TwoBitToSeq(seq.bit, length = 10)
}
\arguments{
\item{seq.bit}{An integer (to be converted to 2-bit)}
}
\value{
A character containing the leetters attro
}
\description{
Convert a given integer to 2-bit and eventually ATCG..
sequence
Based on https://kb.10xgenomics.com/hc/en-us/articles/360004105372-How-do-you-decompress-the-2-bit-barcode-sequences-in-molecule-info-h5-file-
}
