% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{VarianceExplainedPCA}
\alias{VarianceExplainedPCA}
\title{Variance explained in PCA}
\usage{
VarianceExplainedPCA(mtx.genebycell, col_data, factors.of.interest, ...)
}
\description{
Variance explained in PCA
}
