% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{VisualizeWigCoverage}
\alias{VisualizeWigCoverage}
\title{Visualize bigwig coverage for a set of bigwigs over a gene}
\usage{
VisualizeWigCoverage(
  track.data,
  gene.name,
  gtf.data = NULL,
  gtf = NULL,
  exons.gr = NULL,
  tx.ids = NULL,
  ...
)
}
\description{
Visualize bigwig coverage for a set of bigwigs over a gene
}
