% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbmodel.R
\name{EstimateNBPhi}
\alias{EstimateNBPhi}
\title{Estimate phi based on a linear fit to mean-variance relationship}
\usage{
EstimateNBPhi(mean, variance, robust = TRUE)
}
\description{
Estimate phi based on a linear fit to mean-variance relationship
}
