% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{GOClusterGenes}
\alias{GOClusterGenes}
\title{Run clusterProfiler GLM on}
\usage{
GOClusterGenes(df, ont = "BP")
}
\arguments{
\item{df}{Dataframe with columns named gene and celltype}

\item{ont}{Ontology type BP/MF/CC}
}
\description{
Run clusterProfiler GLM on
}
