% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbmodel.R
\name{MeanVarPlot}
\alias{MeanVarPlot}
\title{Plot mean variance relationship for counts or dataframe}
\usage{
MeanVarPlot(
  counts = NULL,
  df = NULL,
  group.by = NULL,
  annotate = FALSE,
  linewidth = 0.75,
  logxy = T
)
}
\description{
Plot mean variance relationship for counts or dataframe
}
