% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{ReadAlevinVelocity}
\alias{ReadAlevinVelocity}
\title{Read data from alevin for celocity analysis creating
two additional assays}
\usage{
ReadAlevinVelocity(gtf.file, frydir)
}
\description{
Read data from alevin for celocity analysis creating
two additional assays
}
